- Shuttle Fleet animation issue fixed - Water rendering cleaned and optimized
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@559 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/VVessel.h
@@ -159,7 +159,7 @@ protected:
 	 * \param mshidx mesh index
 	 * \note If mshidx == (UINT)-1 (default), all meshes are updated.
 	 */
-	void UpdateAnimations (UINT mshidx = (UINT)-1);
+	void UpdateAnimations (int mshidx = -1);
 
 	/**
 	 * \brief Modify local lighting due to planet shadow or
